the project now follows the MVVM, and i made some updates on the onboarding view page to hide the button of the return to the pervious page and make the app follows the test size of the phone system and if it very big the user can scroll down
