--- Splash KAB/Color+Hex.swift
@@ -1,18 +1,0 @@
-import SwiftUI
-
-// Helper to use hex color values
-extension Color {
-    init(hex: String) {
-        let cleanedHex = hex.trimmingCharacters(in: .whitespacesAndNewlines).replacingOccurrences(of: "#", with: "")
-        
-        var rgbValue: UInt64 = 0
-        if Scanner(string: cleanedHex).scanHexInt64(&rgbValue) {
-            let red = Double((rgbValue & 0xFF0000) >> 16) / 255.0
-            let green = Double((rgbValue & 0x00FF00) >> 8) / 255.0
-            let blue = Double(rgbValue & 0x0000FF) / 255.0
-            self.init(red: red, green: green, blue: blue)
-        } else {
-            self.init(red: 0, green: 0, blue: 0) // Fallback to black if invalid hex
-        }
-    }
-}
