--- Splash KAB/ContentViewCam.swift
@@ -1,9 +1,0 @@
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        CameraView()
-            .edgesIgnoringSafeArea(.all)
-    }
-}
