+++ Splash KAB/Utilities/Color+Hex.swift
@@ -1,0 +1,20 @@
+//
+//  Color+Hex.swift
+//  Splash KAB
+//
+//  Created by Shamam Alkafri on 30/12/2024.
+//
+
+import SwiftUI
+
+extension Color {
+    init(hex: String) {
+        let cleanedHex = hex.trimmingCharacters(in: .whitespacesAndNewlines).replacingOccurrences(of: "#", with: "")
+        var rgbValue: UInt64 = 0
+        Scanner(string: cleanedHex).scanHexInt64(&rgbValue)
+        let red = Double((rgbValue & 0xFF0000) >> 16) / 255.0
+        let green = Double((rgbValue & 0x00FF00) >> 8) / 255.0
+        let blue = Double(rgbValue & 0x0000FF) / 255.0
+        self.init(red: red, green: green, blue: blue)
+    }
+}
