+++ Splash KAB/Views/CameraViewController.swift
@@ -1,21 +1,30 @@
-import UIKit
+//
+//  CameraViewController.swift
+//  Splash KAB
+//
+//  Created by Shamam Alkafri on 30/12/2024.
+//
+
+import SwiftUI
 import AVFoundation
+import Foundation
 import Vision
+import UIKit
 
 class CameraViewController: UIViewController {
+    var cameraModel: CameraModel?
     private var captureSession: AVCaptureSession!
     private var previewLayer: AVCaptureVideoPreviewLayer!
     private var speechSynthesizer = AVSpeechSynthesizer()
     private var lastInstructionTime: Date?
-    private var visualFeedbackLayer: CAShapeLayer! // Visual feedback
+    private var visualFeedbackLayer: CAShapeLayer!
 
     override func viewDidLoad() {
         super.viewDidLoad()
         setupCamera()
-        setupVisualFeedback() //  visual feedback
+        setupVisualFeedback()
     }
 
-    /// Sets up the camera feed
     private func setupCamera() {
         captureSession = AVCaptureSession()
         guard let videoCaptureDevice = AVCaptureDevice.default(for: .video) else { return }
@@ -45,7 +54,6 @@ class CameraViewController: UIViewController {
         captureSession.startRunning()
     }
 
-    /// Sets up the visual feedback layer
     private func setupVisualFeedback() {
         visualFeedbackLayer = CAShapeLayer()
         visualFeedbackLayer.strokeColor = UIColor.green.cgColor
@@ -54,14 +62,12 @@ class CameraViewController: UIViewController {
         view.layer.addSublayer(visualFeedbackLayer)
     }
 
-    /// Draws a rectangle for the detected box
     private func drawVisualFeedback(for boundingBox: CGRect) {
         let frameWidth = view.frame.width
         let frameHeight = view.frame.height
 
-        // Scale the detected bounding box to the screen size
         let boxX = boundingBox.minX * frameWidth
-        let boxY = (1 - boundingBox.maxY) * frameHeight // Flip Y-axis
+        let boxY = (1 - boundingBox.maxY) * frameHeight
         let boxWidth = boundingBox.width * frameWidth
         let boxHeight = boundingBox.height * frameHeight
 
@@ -72,17 +78,15 @@ class CameraViewController: UIViewController {
         }
     }
 
-    /// Clears the visual feedback from the screen
     private func clearVisualFeedback() {
         DispatchQueue.main.async {
             self.visualFeedbackLayer.path = nil
         }
     }
 
-    /// Speaks a message to guide the user
     private func speakMessage(_ message: String) {
         if let lastTime = lastInstructionTime, Date().timeIntervalSince(lastTime) < 3 {
-            return // Avoid speaking too frequently
+            return
         }
 
         lastInstructionTime = Date()
@@ -91,16 +95,14 @@ class CameraViewController: UIViewController {
         speechSynthesizer.speak(utterance)
     }
 
-    /// Handles detected box's position and guides the user
     private func handleDetectedBox(boundingBox: CGRect) {
         let frameWidth = view.frame.width
         let frameHeight = view.frame.height
 
-        // Get the edges of the detected box
         let boxMinX = boundingBox.minX * frameWidth
         let boxMaxX = boundingBox.maxX * frameWidth
-        let boxMinY = (1 - boundingBox.maxY) * frameHeight // Flip Y-axis
-        let boxMaxY = (1 - boundingBox.minY) * frameHeight // Flip Y-axis
+        let boxMinY = (1 - boundingBox.maxY) * frameHeight
+        let boxMaxY = (1 - boundingBox.minY) * frameHeight
 
         let boxWidth = boxMaxX - boxMinX
         let boxHeight = boxMaxY - boxMinY
@@ -110,7 +112,6 @@ class CameraViewController: UIViewController {
 
         var needsGuidance = false
 
-        // Check if parts of the box are outside the screen, and provide guidance to the user
         if boxMinX < 0 {
             speakMessage("Move right")
             needsGuidance = true
@@ -127,7 +128,6 @@ class CameraViewController: UIViewController {
             needsGuidance = true
         }
 
-        // Check if the box is too close or too far
         if boxWidth < frameWidth * 0.2 {
             speakMessage("Move the box closer")
             needsGuidance = true
@@ -136,18 +136,15 @@ class CameraViewController: UIViewController {
             needsGuidance = true
         }
 
-        // If the box is within the screen and fully detected, give feedback
         if !needsGuidance && boxWidth > frameWidth * 0.2 && boxHeight > frameHeight * 0.2 {
             speakMessage("Box fully detected")
         }
 
-        // Draw visual feedback for the detected box
         drawVisualFeedback(for: boundingBox)
     }
 }
 
 extension CameraViewController: AVCaptureVideoDataOutputSampleBufferDelegate {
-    /// Processes each frame from the camera to detect a box
     func captureOutput(_ output: AVCaptureOutput, didOutput sampleBuffer: CMSampleBuffer, from connection: AVCaptureConnection) {
         guard let pixelBuffer = CMSampleBufferGetImageBuffer(sampleBuffer) else { return }
         let requestHandler = VNImageRequestHandler(cvPixelBuffer: pixelBuffer, options: [:])
@@ -156,23 +153,19 @@ extension CameraViewController: AVCaptureVideoDataOutputSampleBufferDelegate {
             guard let results = request.results as? [VNRectangleObservation], !results.isEmpty else {
                 self.speakMessage("No box detected")
                 self.clearVisualFeedback()
-                print("No box detected")  // Debugging print statement
                 return
             }
 
-            // Get the largest box by area
             if let largestBox = results.max(by: { $0.boundingBox.width * $0.boundingBox.height < $1.boundingBox.width * $1.boundingBox.height }) {
-                print("Box detected with bounding box: \(largestBox.boundingBox)")  // Debugging print statement
                 self.handleDetectedBox(boundingBox: largestBox.boundingBox)
             }
         }
 
-        // Adjust the rectangle detection parameters
-        request.minimumAspectRatio = 0.5  // Allow more flexibility
-        request.maximumAspectRatio = 2.0  // Allow more flexibility
-        request.minimumSize = 0.2  // 20% size of the screen
-        request.maximumObservations = 5  // Allow more than 1 box
-        request.minimumConfidence = 0.5  // Lower confidence threshold
+        request.minimumAspectRatio = 0.5
+        request.maximumAspectRatio = 2.0
+        request.minimumSize = 0.2
+        request.maximumObservations = 5
+        request.minimumConfidence = 0.5
 
         do {
             try requestHandler.perform([request])
@@ -181,5 +174,3 @@ extension CameraViewController: AVCaptureVideoDataOutputSampleBufferDelegate {
         }
     }
 }
-
-
