--- Splash KAB/myPillsApp.swift
@@ -1,17 +1,0 @@
-//
-//  myPillsApp.swift
-//  myPills
-//
-//  Created by Reeman on 15/12/2024.
-//
-
-import SwiftUI
-
-struct myPillsApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
-
